added SetBranchAddress method

+++ example/thist.jl
@@ -4,7 +4,7 @@ using Base.Test
 #println(methods(TH1D))
 hi = TH1D("my_hist", "My Hist", int32(10), -3.0, 3.0)
 @test bytestring(GetName(hi)) == "my_hist"
-@test GetNbinsX(hi) == 10
+@test GetNbinsX(hi) == int32(10)
 
 n=1000000
 for i=1:n
